Revamped ViewModels to allow for same task card component to be used in both Home and History screens
