+++ app/src/main/java/com/metropolia/eatthefrog/screens/history/components/CompletedTasksContainer.kt
@@ -19,6 +19,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.metropolia.eatthefrog.R
 import com.metropolia.eatthefrog.constants.DATE_FORMAT
+import com.metropolia.eatthefrog.screens.home.components.SingleTaskContainer
 import com.metropolia.eatthefrog.viewmodels.HistoryScreenViewModel
 import java.text.SimpleDateFormat
 
@@ -43,7 +44,7 @@ fun CompletedTasksContainer(vm: HistoryScreenViewModel) {
 
             itemsIndexed(tasks.value) { index, task ->
 
-                if (selectedType.value.contains(task.taskType.name) || selectedType.value.contains(all)) {
+                if (selectedType.value.contains(task.taskTypeOld.name) || selectedType.value.contains(all)) {
 
                     if (index == 0) {
                         Text(task.deadline, fontWeight = FontWeight.Bold, modifier = Modifier.padding(top = 10.dp))
@@ -62,7 +63,7 @@ fun CompletedTasksContainer(vm: HistoryScreenViewModel) {
                         horizontalArrangement = Arrangement.Start,
                         verticalAlignment = Alignment.CenterVertically
                     ) {
-                        HistoryScreenTaskContainer(task, vm)
+                        SingleTaskContainer(task, vm)
                     }
 
                 }
