+++ app/src/main/java/com/metropolia/eatthefrog/ui_components/SingleTaskContainer.kt
@@ -29,6 +29,7 @@ import com.metropolia.eatthefrog.placeholder_data.PlaceholderTask
 import com.metropolia.eatthefrog.placeholder_data.PlaceholderTasks
 import com.metropolia.eatthefrog.viewmodels.DateFilter
 import com.metropolia.eatthefrog.viewmodels.HomeScreenViewModel
+import com.metropolia.eatthefrog.viewmodels.TasksViewModel
 import java.util.*
 
 /**
@@ -36,7 +37,7 @@ import java.util.*
  */
 @OptIn(ExperimentalMaterialApi::class)
 @Composable
-fun SingleTaskContainer(task: Task, vm: HomeScreenViewModel) {
+fun SingleTaskContainer(task: Task, vm: TasksViewModel) {
     val backgroundColor = if (task.isFrog) MaterialTheme.colors.primaryVariant else Color.White
     val taskNameTextColor = if (task.isFrog) Color.White else Color.Black
     val subtaskTextColor = if (task.isFrog) MaterialTheme.colors.secondary else MaterialTheme.colors.primary
@@ -45,7 +46,7 @@ fun SingleTaskContainer(task: Task, vm: HomeScreenViewModel) {
     val closedSubtaskAmount = subtasks.value?.filter { it.completed }?.size ?: 0
     val subtaskText = if (subtaskAmount.value == 0) stringResource(id = R.string.no_subtasks)
                       else "$closedSubtaskAmount/${subtaskAmount.value} ${stringResource(id = if (subtaskAmount.value == 1) R.string.subtask else R.string.subtasks)} ${stringResource(id = R.string.done)}"
-    val typeText = task.taskType.toString().lowercase().capitalize()
+    val typeText = task.taskTypeOld.toString().lowercase().capitalize()
     val deadlineText = if (vm.selectedFilter.value == DateFilter.TODAY) "${stringResource(id = R.string.at)} ${task.time}" else "${task.deadline} ${stringResource(id = R.string.at)} ${task.time}"
 
     Card(
